today we learn about *read-suppress*
As much fun as it may be to learn about conditions and restarts I'm
really happy that *read-suppress* is a thing, because trying to deal
with symbols that have colons in them is a nightmare that I would
rather deal with in clacket than here. Portability is going to be a
mess for that based on the work that led up to this commit.